Move RealThreadPool to internal packaged -LogDispatcher file
