+++ library/file/api/file.api
@@ -17,7 +17,7 @@ public final class io/matthewnelson/kmp/log/file/FileLog : io/matthewnelson/kmp/
 	public final field whitelistDomainNull Z
 	public final field whitelistTag Ljava/util/Set;
 	public final field yieldOn B
-	public synthetic fun <init> (Lio/matthewnelson/kmp/log/Log$Level;Lio/matthewnelson/kmp/log/Log$Level;Ljava/io/File;Ljava/util/List;Ljava/lang/String;JLjava/lang/String;Ljava/lang/String;IZLio/matthewnelson/kmp/log/Log$Level;BJLio/matthewnelson/kmp/log/file/FileLog$RealThreadPool;Lio/matthewnelson/kmp/log/file/FileLog$Formatter;ZLjava/util/Set;Ljava/util/Set;ZLjava/util/Set;ZZLjava/lang/String;Lkotlin/jvm/internal/DefaultConstructorMarker;)V
+	public synthetic fun <init> (Lio/matthewnelson/kmp/log/Log$Level;Lio/matthewnelson/kmp/log/Log$Level;Ljava/io/File;Ljava/util/List;Ljava/lang/String;JLjava/lang/String;Ljava/lang/String;IZLio/matthewnelson/kmp/log/Log$Level;BJLio/matthewnelson/kmp/log/file/internal/RealThreadPool;Lio/matthewnelson/kmp/log/file/FileLog$Formatter;ZLjava/util/Set;Ljava/util/Set;ZLjava/util/Set;ZZLjava/lang/String;Lkotlin/jvm/internal/DefaultConstructorMarker;)V
 	public final fun getBufferOverflowDropOldest ()Z
 	public final fun isActive ()Z
 	public final fun pendingLogCount ()J
@@ -80,7 +80,6 @@ public abstract interface class io/matthewnelson/kmp/log/file/FileLog$Formatter 
 public abstract class io/matthewnelson/kmp/log/file/FileLog$ThreadPool {
 	public static final field Companion Lio/matthewnelson/kmp/log/file/FileLog$ThreadPool$Companion;
 	public final field nThreads I
-	public synthetic fun <init> (ILkotlin/jvm/internal/DefaultConstructorMarker;)V
 	public static final fun of (I)Lio/matthewnelson/kmp/log/file/FileLog$ThreadPool;
 	public fun toString ()Ljava/lang/String;
 }
